pyproject.toml
[tool.poetry]
name = "infotecs"
version = "0.1.0"
description = ""
authors = ["ledxdeliveryflopp <korstim18@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
pytest = "^8.3.2"
pydantic = "^2.8.2"
pydantic-settings = "^2.4.0"
httpx = "^0.27.0"
pytest-asyncio = "^0.23.8"
allure-pytest = "^2.13.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = [".", "src"]